+++ packages/plugin/convert/src/_shared/main.ts
@@ -1,6 +1,7 @@
 
 import {
 	ensureDir,
+	getRandomUUID,
 	getStringsFrom,
 	getTempDir,
 	joinPath,
@@ -30,6 +31,14 @@ export class ConvertSuper<Props extends ConvertPropsSuper> {
 
 	}
 
+	protected async _forEachContent( input: string[] | string, cb: ( data: Awaited<ReturnType<typeof getStringsFrom>>[number] ) => Promise<void> ) {
+
+		const inputs = await this._getContent( input )
+
+		await Promise.all( inputs.map( async i => await cb( i ) ) )
+
+	}
+
 	protected async _writeOutput( out: string, id: string, content: string ) {
 
 		await ensureDir( out )
@@ -43,7 +52,7 @@ export class ConvertSuper<Props extends ConvertPropsSuper> {
 	protected async _getOutput( ) {
 
 		// import be custom folder inside tempr dir, for then remove it
-		const tempDir = joinPath( getTempDir(), 'dovenv-convert' )
+		const tempDir = joinPath( getTempDir(), 'dovenv-convert', getRandomUUID() )
 
 		const dir = this.props.output ? this.props.output : tempDir