+++ packages/plugin/convert/src/_shared/types.ts
@@ -32,3 +32,4 @@ export type ConvertResponse = {
 }[]
 
 export type ConvertSuperInterface = { run: ( ) => Promise<ConvertResponse> }
+