+++ packages/utils-media/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dovenv/utils-media
 
+## 1.5.5
+
+### Patch Changes
+
+- Fix predocs issues, add docs
+
+- Updated dependencies []:
+  - @dovenv/utils@1.5.5
+
 ## 1.5.4
 
 ### Patch Changes